requirements.txt
flask
speedtest-cli